aggiunto la parte dei punti al db

+++ src/code/domain_classes/F1_Item.java
@@ -9,9 +9,9 @@ public interface F1_Item {
 
     void setFantaF1points(int FantaF1points);
 
-    int getFantavalue();
+    float getFantavalue();
 
-    void setFantavalue(int Fantavalue);
+    void setFantavalue(float Fantavalue);
 
     void setName(String name);
